api/requirements.txt
Django
djangorestframework
gunicorn
django-storages
boto3
django-cors-headers
